Created CustomTabPanel component
