+++ frontend/src/components/tabs/CustomTabPanel.tsx
@@ -1,0 +1,33 @@
+import { Box, Typography } from '@mui/material';
+
+interface TabPanelProps extends React.HTMLAttributes<HTMLDivElement> {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+}
+
+export const CustomTabPanel = (props: TabPanelProps) => {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      {...other}
+    >
+      {value === index && (
+        <Box sx={{ p: 3 }}>
+          <Typography>{children}</Typography>
+        </Box>
+      )}
+    </div>
+  );
+};
+
+CustomTabPanel.defaultProps = {
+  children: '',
+};
+
+export default CustomTabPanel;
